Allow decimal place for quantities

+++ CompanyStudio/Models/LocationItem.cs
@@ -13,7 +13,7 @@ namespace CompanyStudio.Models
         public Location Location { get; set; }
         public long? ItemID { get; set; }
         public Item Item { get; set; }
-        public short? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
         public decimal? BasePrice { get; set; }
         public PromotionLocationItem CurrentPromotionLocationItem { get; set; }
